+++ packages/components/IconPicker/IconPickerDemo.tsx
@@ -12,6 +12,7 @@ const useStyles = createStyles(({ token }) => ({
     display: 'flex',
     flexDirection: 'column',
     backgroundColor: token.colorBgBase,
+    border: `1px solid ${token.colorBorderSecondary}`,
   },
   header: {
     'backgroundColor': token.colorBgBase,